refactor(cli): extract undo_apply atomic tests into sibling module (#609 phase 3)
`crates/cli/src/cli/commands/undo_apply.rs` (3944 LOC) bundled the
undo/redo apply command surface with a 1766-LOC trailing
`#[cfg(test)] mod atomic_tests`. Convert the file to a directory module
(`undo_apply/mod.rs`) and move the test body verbatim into a sibling
`undo_apply/atomic_tests.rs`, declared via `#[cfg(test)] mod atomic_tests;`.

Pure, behavior-preserving re-modularization: the test body moved unchanged
(de-indented one level), still references the parent via `super::*`, and the
module keeps its original name. No logic, signature, or public-surface change
-- git sees the test block as a move out of mod.rs into the sibling file.

Same seam as #609 phases 1/2 (command_catalog tests #815, git_overlay_health
#816). Phases 4-6 (catalog generator/data split, repository.rs, mount/core.rs)
remain.
